requirements/common.txt
Django==5.1.3
psycopg==3.2.3
psycopg2-binary==2.9.10
django-allauth==65.2.0
django-tenants==3.7.0
djangorestframework==3.15.2
